requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
Pillow==10.1.0
razorpay==1.4.1
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.3
requests==2.31.0
mysql-connector-python==8.0.33
firebase-admin==7.3.0
